🐛 fix: 어드민 페이지 route 컴포넌트로 감싸는 방식으로 수정
