+++ src/pages/admin/comments.tsx
@@ -1,12 +1,9 @@
-import React, { useEffect } from 'react';
-import { useRouter } from 'next/router';
+import React from 'react';
+
 import { useMutation, useQueryClient } from 'react-query';
-import { useRecoilValue } from 'recoil';
 
 import { useCommentListQuery } from '../../api/hook/admin';
 import { deleteComment } from '../../api/fetcher/admin';
-import { userIdState } from '../../recoils/atoms/users';
-import { useUserQuery } from '../../api/hook/users';
 
 import HomeLayout from '../../components/common/layout/layout';
 import { Container, Title, Section, Tr, Td, PageContent } from '../../styles/admin/user';
@@ -14,21 +11,11 @@ import Button from '../../components/common/button';
 import ArrowLeftIcon from '../../components/common/@Icons/arrowLeftIcon';
 import ArrowRightIcon from '../../components/common/@Icons/arrowRightIcon';
 import GarbageIcon from '../../components/common/@Icons/garbageIcon';
+import AdminRoute from '../../components/common/routes/admin';
 
 const Comments = () => {
   const [page, setPage] = React.useState<number>(1);
   const [totalPage, setTotalPage] = React.useState<number>(0);
-  const router = useRouter();
-  const userId = useRecoilValue(userIdState);
-
-  useUserQuery(userId as number, {
-    onSuccess: data => {
-      if (!data.isAdmin) {
-        router.replace('/');
-        return;
-      }
-    },
-  });
 
   const queryClient = useQueryClient();
   const { data: commentList } = useCommentListQuery({
@@ -46,10 +33,6 @@ const Comments = () => {
       queryClient.invalidateQueries(['COMMENTSLIST', page]);
     },
   });
-
-  useEffect(() => {
-    if (!userId) router.replace('/');
-  }, []);
 
   return (
     <Container>
@@ -93,5 +76,9 @@ const Comments = () => {
 export default Comments;
 
 Comments.getLayout = function getLayout(page: React.ReactElement) {
-  return <HomeLayout>{page}</HomeLayout>;
+  return (
+    <HomeLayout>
+      <AdminRoute>{page} </AdminRoute>
+    </HomeLayout>
+  );
 };
