+++ src/pages/admin/index.tsx
@@ -1,4 +1,4 @@
-import React, { useEffect } from 'react';
+import React from 'react';
 import { useRouter } from 'next/router';
 
 import { useCommentListQuery, useUserListQuery } from '../../api/hook/admin';
@@ -8,6 +8,7 @@ import HomeLayout from '../../components/common/layout/layout';
 import ArrowRightIcon from '../../components/common/@Icons/arrowRightIcon';
 import { AdminContainer, Title, UnderLine, Section, Article, ArticleTitle } from '../../styles/admin';
 import { Td, Tr } from '../../styles/admin/user';
+import AdminRoute from '../../components/common/routes/admin';
 
 const Admin = () => {
   const router = useRouter();
@@ -78,5 +79,9 @@ const Admin = () => {
 export default Admin;
 
 Admin.getLayout = function getLayout(page: React.ReactElement) {
-  return <HomeLayout>{page}</HomeLayout>;
+  return (
+    <HomeLayout>
+      <AdminRoute>{page} </AdminRoute>
+    </HomeLayout>
+  );
 };
